% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{function for MALDI file cropping
 Imports: Cardinal
 @file - MALDI file name
 @xlim - x axes of
 example: cerebellum.crop <- crop(cer, xlim=c(50,100), ylim=c(150,200))}
\usage{
crop(file, xlim, ylim)
}
\description{
function for MALDI file cropping
 Imports: Cardinal
 @file - MALDI file name
 @xlim - x axes of
 example: cerebellum.crop <- crop(cer, xlim=c(50,100), ylim=c(150,200))
}
