% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.spec.R
\name{extract.spec}
\alias{extract.spec}
\title{Function for extracting spectrum into a table for MALDI files
 Imports: Cardinal
 @file - MALDI file name
 example: cer.spec <- extract.spec(cer)}
\usage{
extract.spec(file)
}
\description{
Function for extracting spectrum into a table for MALDI files
 Imports: Cardinal
 @file - MALDI file name
 example: cer.spec <- extract.spec(cer)
}
