% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz.annotation.R
\name{mz.annotation}
\alias{mz.annotation}
\title{function mz.annotation by mz
 This function uses a string of mz values as an input and generates a list of putative lipids with three most common adducts
(H+, Na+, K+) by sendig automated requests to lipidmaps.org and parsing the output data into a table
 Imports: RCurl
 Imports: stringr
 @mzlist - a string of mz values to be annotated
 example: mz.values <- c(159.2841, 273.3309, 966.4836)
 annot.table <- mz.annotation(mz.values)}
\usage{
mz.annotation(mzlist)
}
\description{
function mz.annotation by mz
 This function uses a string of mz values as an input and generates a list of putative lipids with three most common adducts
(H+, Na+, K+) by sendig automated requests to lipidmaps.org and parsing the output data into a table
 Imports: RCurl
 Imports: stringr
 @mzlist - a string of mz values to be annotated
 example: mz.values <- c(159.2841, 273.3309, 966.4836)
 annot.table <- mz.annotation(mz.values)
}
