% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processing}
\alias{processing}
\title{function combining all MALDI processing steps in one
 Imports: Cardinal
 @file - MALDI file name
 @reference - reference mzlist for alignment
 example: cer.proc <- processing(cer, mz(cer2))}
\usage{
processing(file, reference)
}
\description{
function combining all MALDI processing steps in one
 Imports: Cardinal
 @file - MALDI file name
 @reference - reference mzlist for alignment
 example: cer.proc <- processing(cer, mz(cer2))
}
