% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.matr.R
\name{remove.matr}
\alias{remove.matr}
\title{Function for extracting matrix peaks from the sample.
 Returns a cropped MALDI file cleaned from the peaks that are more intensive in the matrix
 Imports: Cardinal
 Imports: tidyr
 @file - MALDI file name
 @file.skmg - SpatialKMeans2 file for the same dataset as @file
 @s - number(s) of cluster(s) referring to the sample
 @threshold - threshold for matrix filtering
 example: cer.cl <- remove.matr(cer, cer.skmg, s = c(2,3), threshold = 1.5)}
\usage{
remove.matr(file, file.skmg, s, threshold)
}
\description{
Function for extracting matrix peaks from the sample.
 Returns a cropped MALDI file cleaned from the peaks that are more intensive in the matrix
 Imports: Cardinal
 Imports: tidyr
 @file - MALDI file name
 @file.skmg - SpatialKMeans2 file for the same dataset as @file
 @s - number(s) of cluster(s) referring to the sample
 @threshold - threshold for matrix filtering
 example: cer.cl <- remove.matr(cer, cer.skmg, s = c(2,3), threshold = 1.5)
}
